Split thumbnail tests up into multiple files Added conditions for isAdmin, isGroupOwner in PermissionsTest
